Infrastructure checkpoint

+++ src/Organizr.Application/TodoLists/Commands/DeleteTodoItem/DeleteTodoItemCommand.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Ardalis.GuardClauses;
 using MediatR;
 using Organizr.Application.Common.Exceptions;
+using Organizr.Application.Common.Interfaces;
 using Organizr.Domain.Lists.Entities.TodoListAggregate;
 
 namespace Organizr.Application.TodoLists.Commands.DeleteTodoItem
@@ -22,12 +23,19 @@ namespace Organizr.Application.TodoLists.Commands.DeleteTodoItem
 
     public class DeleteTodoItemCommandHandler : IRequestHandler<DeleteTodoItemCommand>
     {
+        private readonly ICurrentUserService _currentUserService;
+        private readonly IResourceAccessService _resourceAccessService;
         private readonly ITodoListRepository _todoListRepository;
 
-        public DeleteTodoItemCommandHandler(ITodoListRepository todoListRepository)
+        public DeleteTodoItemCommandHandler(ICurrentUserService currentUserService,
+            IResourceAccessService resourceAccessService, ITodoListRepository todoListRepository)
         {
+            Guard.Against.Null(currentUserService, nameof(currentUserService));
+            Guard.Against.Null(resourceAccessService, nameof(resourceAccessService));
             Guard.Against.Null(todoListRepository, nameof(todoListRepository));
-            
+
+            _currentUserService = currentUserService;
+            _resourceAccessService = resourceAccessService;
             _todoListRepository = todoListRepository;
         }
 
@@ -37,6 +45,9 @@ namespace Organizr.Application.TodoLists.Commands.DeleteTodoItem
 
             if (todoList == null)
                 throw new NotFoundException<TodoList>(request.TodoListId);
+
+            if (!_resourceAccessService.CanAccess(request.TodoListId, _currentUserService.UserId))
+                throw new AccessDeniedException(request.TodoListId, _currentUserService.UserId);
 
             todoList.DeleteTodo(request.Id);
